Split the "/health" endpoint into 3.
This allows us to differentiate between "readiness", "liveness", and
"connectedness".

The first two do not require talking to external services, whereas
"connectedness" does. As such, "readiness" and "liveness" are more
appropriate for automated checks (for example, in Docker or Kubernetes),
and "connectedness" is more useful when diagnosing an issue.

In addition, the health check endpoints now return a structured error.